break CreatePost in smaller and reusable components (CreatePostModal)
