+++ src/components/kanban/CreatePostModal.js
@@ -1,0 +1,149 @@
+import React, { useState, useEffect } from 'react';
+import { useDispatch, useSelector } from 'react-redux';
+import { addList, addPost } from '../../actions/listsActions';
+import {
+    makeStyles,
+    Dialog,
+    DialogTitle,
+    DialogContent,
+    DialogActions,
+    TextField,
+    FormControl,
+    InputLabel,
+    Select,
+    MenuItem,
+    Button
+} from '@material-ui/core';
+
+const useStyles = makeStyles(theme => ({
+    content: {
+      width: '400px',
+      maxWidth: '100%'
+    },
+    formControl: {
+      minWidth: 160
+    }
+  }));
+
+const CreatePostModal = (props) => {
+    const {
+        listId,
+        open,
+        handleClose,
+    } = props;
+
+    const dispatch = useDispatch();
+    const { lists } = useSelector(state => state.kanban);
+    const [post, setPost] = useState({
+        list_id: "",
+        post_text: ""
+    });
+
+    const { content, formControl } = useStyles();
+    
+    useEffect(() => {
+        if (lists) {
+            const drafts = Object.values(lists).find(list => list.title === 'Drafts');
+    
+            if (!drafts) {
+                // create Drafts list
+                dispatch(addList('Drafts'));
+            } else {
+                // set Drafts or listId from param as default list
+                setPost(prevPost => ({
+                ...prevPost,
+                list_id: listId ? listId : drafts.id
+                }));
+            }
+        }
+    }, [lists, listId, dispatch]);
+
+    const handlePostText = (e) => {
+        const target = e.currentTarget;
+        setPost(prevPost => ({
+            ...prevPost,
+            [target.id]: target.value
+        }));
+    }
+
+    const handleSelectList = (e) => {
+        const target = e.target;
+
+        setPost(prevPost => ({
+            ...prevPost,
+            list_id: target.value
+        }));  
+    }
+
+    const handleCreatePost = (e) => {
+        e.preventDefault();
+        dispatch(addPost(post));
+        handleClose();
+    }
+    
+    return (
+        <Dialog
+        open={open}
+        onClose={handleClose}
+        aria-labelledby="form-dialog-title"
+      >
+        <DialogTitle id="form-dialog-title">Create a Post</DialogTitle>
+        <DialogContent className={content}>
+          {/* Post text */}
+          <TextField
+            autoFocus
+            margin="dense"
+            id="post_text"
+            label="Post text"
+            fullWidth
+            multiline
+            rows={4}
+            variant="outlined"
+            onChange={handlePostText}
+          />
+
+        {/* List Select */}
+        <FormControl className={formControl}>
+          <InputLabel id="list_label">Category</InputLabel>
+          <Select
+            labelId="list_label"
+            id="list_select"
+            value={post.list_id}
+            onChange={handleSelectList}
+          >
+            {lists && Object.values(lists).map(list => {
+              return (
+                <MenuItem key={list.id} value={list.id}>{list.title}</MenuItem>
+              )
+            })}
+          </Select>
+        </FormControl>
+
+          {/* Image upload */}
+          {/* <input
+            accept="image/*"
+            style={{ display: "none" }}
+            id="create-tweet"
+            multiple
+            type="file"
+          />
+          <label htmlFor="create-tweet">
+            <AddAPhotoIcon />
+          </label> */}
+        </DialogContent>
+
+        <DialogActions
+          style={{ display: "flex", justifyContent: "space-between" }}
+        >
+          <Button onClick={handleClose} color="primary">
+            Cancel
+          </Button>
+          <Button onClick={handleCreatePost} color="primary">
+            Add to List
+          </Button>
+        </DialogActions>
+      </Dialog>
+    )
+}
+
+export default CreatePostModal
